requirements.txt
google-genai
requests
pydantic
openai
